+++ event-ticket-api/src/main/java/com/example/booking/dto/TicketCategoryDto.java
@@ -1,4 +1,6 @@
 package com.example.booking.dto;
 
-public record TicketCategoryDto(String name, Double price, Integer availableTicketsFromCategory) {
+import java.math.BigDecimal;
+
+public record TicketCategoryDto(String name, BigDecimal price, Integer availableTicketsFromCategory) {
 }
